requirements.txt
django
djangoajax
django-background-tasks
PyYAML
netifaces
docker
python-gvm
xmltodict